server/requirements.txt
aiosqlite==0.21.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
bcrypt==3.2.2
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.120.2
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.3.4
openpyxl==3.1.5
pandas==2.3.3
passlib==1.7.4
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.2
sentry-sdk==2.43.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
